% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_tools.r
\name{akimafun}
\alias{akimafun}
\title{univariate Akima interpolation}
\usage{
akimafun(x, y)
}
\arguments{
\item{x}{Independent variable data}

\item{y}{Dependent variable data}
}
\value{
An interpolating function, in the spirit of \link{stats::splinefun}.
}
\description{
Creates an interpolating function of irregular, univariate data.
}
\details{
Employs Akima's algorithm for univariate interpolation. This is a Hermite
interpolation method, prescribing slopes at data points depending on the
neighbouring points. It is a local method.

Adapted from the "akimaInterp" function in the pracma package, written by Hans W. Borchers.
(Which is also licensed under GPLv3.)
}
\references{
Akima, Hiroshi, "A New Method of Interpolation and Smooth Curve
Fitting Based on Local Procedures", 1970.
}
