% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_coding.r
\name{alternative_rate_function}
\alias{alternative_rate_function}
\title{This implements the rate coding model proposed in
Petersen and Rostalski (2019): A Comprehensive Mathematical Model of Motor Unit Pool Organization, Surface
Electromyography, and Force Generation. frontiers in Physiology. https://doi.org/10.3389/fphys.2019.00176}
\usage{
alternative_rate_function(muscle_load, params)
}
\description{
This implements the rate coding model proposed in
Petersen and Rostalski (2019): A Comprehensive Mathematical Model of Motor Unit Pool Organization, Surface
Electromyography, and Force Generation. frontiers in Physiology. https://doi.org/10.3389/fphys.2019.00176
}
