% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.r
\name{apply_electrode_configs}
\alias{apply_electrode_configs}
\title{Apply electrode configurations}
\usage{
apply_electrode_configs(muscle_potentials, electrode_configs)
}
\arguments{
\item{muscle_potentials}{A \code{data.frame} as usually created by a call to
\code{\link{sum_muscle_contribs}}, containing the contributions of
different muscles to the monopolar signals measured at each electrode.}

\item{electrode_configs}{A list of vectors, where each vector should have as
many elements as there are electrodes. The elements of the vector
determine the weights given to the different monopolar signals for
calculating the combined signal.}
}
\value{
A \code{data.frame} containing the contributions of the different
  muscles to the signals obtained using the different electrode
  configurations. These are still specified via the categorical variable
  \code{electrode} (instead of using "electrode_config", or something
  similar), in order to allow for easy usage of the same analysis functions.
}
\description{
Calculate combinations of different, monopolar electrode signals.
}
\details{
This allows for the implementation of differential measurements, etc.
}
\note{
This implementation is probably highly inefficient in every regard.
}
