% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_transformation.r
\name{calc_activation_transformations}
\alias{calc_activation_transformations}
\title{Calculate nonlinear transformations of activations signals}
\usage{
calc_activation_transformations(MUs, MUs_rate_functions)
}
\arguments{
\item{MUs}{A data.frame of MUs belonging to (potentially) multiple muscles.}

\item{MUs_rate_functions}{A data frame containing a column $rate.function that gives the
firing rate as a function of muscle activation for each MU.}
}
\value{
A list of functions that gives the muscle input as a function of the desired
relative output force for each muscle.
}
\description{
Calculate a nonlinear input transformation for each muscle that makes the
muscle generated the desired relative output force.
}
\details{
Depending on the firing rate, MU pool, and twitch force model, a nonlinear
transformation of the muscle activation input is calculated, such that the
generated relative output muscle force matches the input activation signal in
the mean over time.
}
