% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{calc_affine_param}
\alias{calc_affine_param}
\title{Calculate a variable that is affine in another variable, given that
variable's value.}
\usage{
calc_affine_param(min_val, max_val, ratio)
}
\arguments{
\item{min_val}{Minimum value of the variable to calculate}

\item{max_val}{Maximum value of the variable to calculate}

\item{ratio}{Value between 0 and 1, indicating position between min and max}
}
\value{
(Scalar) value of the affine dependent parameter
}
\description{
Calculates
`var2 = var2min + (var1i - var1min) / (var1max - var1min) * (var2max - var2min)`
where the quotient `(var1i - var1min) / (var1max - var1min)` corresponds to
function parameter `ratio`.
}
