% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_coding.r
\name{calc_MU_firing_rate_funcs}
\alias{calc_MU_firing_rate_funcs}
\title{Calculate the firing-rate-force relationship of all MUs}
\usage{
calc_MU_firing_rate_funcs(muscles, MUs)
}
\arguments{
\item{muscles}{A data frame containing information on the muscles.}

\item{MUs}{A data frame containing MU objects.}
}
\value{
A data frame containing a column $rate.function that gives the
  function described above for each MU.
}
\description{
For each MU calculate a function that gives its firing rate in dependence of
the current muscle work load.
}
\details{
Note that this includes the different MU's activation threshold, as before
being activated the firing rate of a MU is just zero. The muscle work load
should be given in the ratio of the maximum voluntary contraction (MVC) of
the muscle, i.e., a value between 0 and 1.
}
