% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MU_firing_responses.r
\name{calc_MU_firing_response_fftvals}
\alias{calc_MU_firing_response_fftvals}
\title{Calculate the values of the FFT of the firing response of a MU at an
electrode}
\usage{
calc_MU_firing_response_fftvals(MU_electrode_TF, MU_obj, psi_trafo, sampling)
}
\arguments{
\item{MU_electrode_TF}{A complex vector containing the value of the transfer
function from MU to electrode at frequencies \code{sampling$freqs_to_calc}}

\item{MU_obj}{A MU object.}

\item{psi_trafo}{A function returning the values of the unitary Fourier
transform of psi, given a vector of ordinary input frequencies.}

\item{sampling}{A sampling object as created by \link{setup_sampling}.}
}
\value{
A complex vector that stores the FFT values of the firing response
of MU at electrode.
}
\description{
Take information on the transfer function MU-electrode and information on
the excitatory input signal fed into the MU and combine this to attain the
firing response of the given MU at the given electrode.
}
\details{
This does not involve calculating an FFT! Actually, it is just
\itemize{
\item{1.}{Evaluate the Fourier transform of the excitatory input signal
(given in analytical form) at the same frequencies at which the
MU-electrode TF has been computed.}
\item{2.}{Multiply the complex conjugates of these values with the
corresponding TF values.}
\item{3.}{Obtain the second half of the FFT by mirroring the result,
exploiting the hermitianness of the FFT of a real signal (which
we expect the resulting firing response to be).}
}
}
