% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MU_firing_responses.r
\name{calc_MU_firing_response_fftvals_wrapper}
\alias{calc_MU_firing_response_fftvals_wrapper}
\title{Simple worker function wrapper to facilitate progression display}
\usage{
calc_MU_firing_response_fftvals_wrapper(x, p, ...)
}
\description{
Simple worker function wrapper to facilitate progression display
}
