% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MU_firing_responses.r
\name{calc_MU_firing_responses_fftvals}
\alias{calc_MU_firing_responses_fftvals}
\title{Calculate the values of the FFT of the firing responses of all MUs at all
electrodes}
\usage{
calc_MU_firing_responses_fftvals(
  MUs,
  TFs_MU_input_to_surf_potential,
  psi_trafo,
  num_cores = 1,
  sampling
)
}
\arguments{
\item{MUs}{A data.frame containing a column 'MU.obj' containing MU objects
and index columns 'MU' and 'muscle'.}

\item{TFs_MU_input_to_surf_potential}{A data.frame containing the values of
the transfer functions at frequencies \code{sampling$freqs_to_calc} in
column 'TF', represented as vectors of complex values.}

\item{psi_trafo}{A function returning the values of the unitary Fourier
transform of psi, given a vector of ordinary input frequencies.}

\item{num_cores}{The number of cores that shall be used for computations. If
equal to one (default), calculations are executed sequentially as usual.
Otherwise, the parallel package is used for multi-process computations.}

\item{sampling}{A sampling object as created by \link{setup_sampling}.}
}
\value{
A data.frame containing index columns 'muscle', 'MU' and 'electrode'
as well as a data column 'firing_response_fftvals' that contains vectors
that store the FFT values of the firing response of the particular MU at
the particular electrode.
}
\description{
This is just a wrapper that calls \link{calc_MU_firing_response_fftvals}
for all MU-electrode pairs.
}
