% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.r
\name{calc_shifted_emg_response}
\alias{calc_shifted_emg_response}
\title{Time-shift a signal given by its FFT and compute the IFFT.}
\usage{
calc_shifted_emg_response(firing_response_fft, time_shift, sampling)
}
\arguments{
\item{firing_response_fft}{FFT values of a signal, computed at the
frequencies given by sampling$fft_freqs.}

\item{time_shift}{Amount of time by which the signal should be shifted in
time. Must be zero or positive and smaller than one sampling interval.}

\item{sampling}{An object providing information on the sampling paramters
chosen. Usually constructed by the function create_sampling.}
}
\value{
A real, numerical vector of length sampling$NFFT.
}
\description{
Take FFT values of a signal, perform a small time-shift on it and
compute the inverse FFT.
}
\note{
Requires the underlying signal to be purely real and thus the FFT to be
hermitian.
}
\examples{
set.seed(1)
t <- seq(0.1, 1, 0.1)
y <- sin(2*pi*t)
resp <- semg.simulatr:::calc_shifted_emg_response(
    firing_response_fft = fft(y) / length(y), 1/20,
    sampling = semg.simulatr:::create_sampling(Fs = 10, NFFT = 10))
dev.new()
plot(t, resp)
points(t, y, col="red")

}
