% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{calculate_semg_TFs}
\alias{calculate_semg_TFs}
\title{Calculate transfer functions}
\usage{
calculate_semg_TFs(
  model,
  deterministic = FALSE,
  out_file = NULL,
  num_cores = future::availableCores(omit = 1)
)
}
\description{
Calculate transfer functions
}
\note{
The config file may "link" \code{\link[=calculate_TFs_Far_Mer]{calculate_TFs_Far_Mer()}} to
\code{calc_MU_electrode_time_TFs()} which is called from below.
}
\examples{
model <- calculate_semg_TFs(setup_muscles("config/integration.r", deterministic = TRUE), deterministic = TRUE) 
}
