% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_semg_TFs.r
\name{calculate_TF_Far_Mer_MU_wrapper}
\alias{calculate_TF_Far_Mer_MU_wrapper}
\title{Simple worker function wrapper to facilitate progression display}
\usage{
calculate_TF_Far_Mer_MU_wrapper(x, p, ...)
}
\description{
Simple worker function wrapper to facilitate progression display
}
