% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_semg_TFs.r
\name{calculate_TF_point_Far_Mer}
\alias{calculate_TF_point_Far_Mer}
\title{Evaluate a muscle fiber - electrode TF at a single angular (!) frequency}
\usage{
calculate_TF_point_Far_Mer(freq, electrode, fiber, vol_conductor, Bfun)
}
\description{
Note that this implementation assumes an even, real electrode transfer
function!
}
