% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_semg_TFs.r
\name{calculate_TFs_Far_Mer}
\alias{calculate_TFs_Far_Mer}
\title{Calculate MU - electrode time transfer functions by a precise model}
\usage{
calculate_TFs_Far_Mer(
  MUs,
  electrodes,
  vol_conductor,
  freqs,
  num_cores = 1,
  B_sampling = FALSE
)
}
\arguments{
\item{MUs}{A data.frame containing a column 'MU.obj' containing MU objects
and index columns 'MU' and 'muscle'.}

\item{electrodes}{A data.frame containing columns 'electrode' with the
electrode ID and 'electrode.obj' containing the actual object.}

\item{vol_conductor}{A volume conductor object. Specifies the properties
(i.e. thickness, conductivity) of the tissue layers separating muscle
fibers and electrodes.}

\item{freqs}{A numerical vector containing the (non-angular) frequencies at
which the transfer function is to be evaluated. Results will be returned in
order.
Usually should be generated by a call to the \code{setup.sampling} function
and passing \code{sampling$freqs.to.calc}.}

\item{num_cores}{The number of cores that shall be used for computations. If
equal to one (default), calculations are executed sequentially as usual.
Otherwise, the parallel package is used for multi-process computations.}

\item{B_sampling}{}
}
\value{
A data.frame containing the values of the transfer functions at
frequencies \code{.freqs} in column 'TF', represented as vectors of complex
values.
}
\description{
Compute the transfer functions between MUs and electrodes according to the
model published by Farina and Merletti, https://doi.org/10.1109/10.923782
}
\details{
This calculates the values of the MU-electrode transfer functions for a set
of Motor Units, sampled at a given set of frequencies. Note that the
potentially differing conduction velocities of the MUs will be taken into
account here, s.t. they do neither need to be taken care of in the
specification of .freqs, nor later on in the processing stage.
}
\note{
The config file may "link" the current function to
\code{calc_MU_electrode_time_TFs()} which is called from \code{\link[=calculate_semg_TFs]{calculate_semg_TFs()}}.
}
