% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.r
\name{chunk_MU_contribs}
\alias{chunk_MU_contribs}
\title{Wrapper calling \code{\link[=sample_MU_contribs]{sample_MU_contribs()}} across time axis.}
\usage{
chunk_MU_contribs(
  MU_obj,
  sim_state,
  force,
  excitation,
  rate_function,
  firing_response_fftvals,
  sampling
)
}
\arguments{
\item{MU_obj}{A motor unit object containing information on force twitch
characteristics.}

\item{sim_state}{A list containing elements \code{time_since_last_firing},
\code{Z_score}, \code{contribs} and \code{future_contribs}. Describes the current state
of the simulation of this MU, i.e., summarizes the simulation history.}

\item{force}{Vector of time dependent targeted muscle force values.}

\item{excitation}{Vector of time dependent muscle activation.}

\item{rate_function}{The rate coding function of this MU.}

\item{firing_response_fftvals}{A list of num_electrodes vectors containing
firing response fft values. The index of each vector indicates the
corresponding electrode ID.}

\item{sampling}{An object providing information on the sampling parameters
chosen. Usually constructed by the function create_sampling.}
}
\value{
A list containing the updated \code{sim_state} of this MU and the
generated time dependent \code{emg} and \code{force} values.
}
\description{
Wrapper calling \code{\link[=sample_MU_contribs]{sample_MU_contribs()}} across time axis.
}
