% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{convert_muscle_list_to_df}
\alias{convert_muscle_list_to_df}
\title{Initialize the set of muscles to be simulated}
\usage{
convert_muscle_list_to_df(muscles)
}
\arguments{
\item{muscles}{A list of muscle objects.}
}
\value{
A data.frame containing a column 'muscle' giving the indices of the
  different muscles and a column 'muscle.obj' containing the corresponding
  muscle objects.
}
\description{
Convert data given in the config file to a data.frame.
}
