% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.r
\name{create_planar_3layer_volume_conductor}
\alias{create_planar_3layer_volume_conductor}
\title{Implements the planar three-layer volume conductor model of Farina and Rainoldi (1999),
https://doi.org/10.1016/S1350-4533(99)00075-2}
\usage{
create_planar_3layer_volume_conductor(
  fat_thickness,
  skin_thickness,
  fat_cond,
  skin_cond,
  muscle_cond_fiber,
  muscle_cond_nofiber
)
}
\description{
Implements the planar three-layer volume conductor model of Farina and Rainoldi (1999),
https://doi.org/10.1016/S1350-4533(99)00075-2
}
