% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.r
\name{create_point_electrodes}
\alias{create_point_electrodes}
\title{Create a set of point electrodes}
\usage{
create_point_electrodes(...)
}
\arguments{
\item{...}{Arbitrarily many 2D numerical vectors giving the desired electrode
positions in 3D space with the y coordinate being automatically defined by
the fact that the electrode is placed on the patient's skin. Positions are
given as (x, z).}
}
\value{
A \code{data.frame} containing one row for each electrode. Columns
  are 'electrode' for the id of the electrode and 'electrode.obj' containing
  the actual object of type 'electrode.point'.
}
\description{
Take an arbitrary number of positions on the 2D skin plane and return a data
frame containing point electrodes positioned there.
}
