% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_coding.r
\name{create_rate_function_creators}
\alias{create_rate_function_creators}
\title{Set up rate coding for a set of muscles}
\usage{
create_rate_function_creators(muscles)
}
\arguments{
\item{muscles}{A data frame containing the muscle objects for which the rate
coding shall be set up.}
}
\value{
A data frame containing for each muscle a function that takes the
  recruitment threshold of a MU in that muscle and returns a function giving
  the firing rate of that MU in dependence of the current muscle work load.
}
\description{
This wraps \link{create_rate_function_creator} by applying it to a data frame
of muscles and coercing the output into a data frame again.
}
