% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{distribute_MUs_in_muscle}
\alias{distribute_MUs_in_muscle}
\title{Place MUs somewhere in the muscle area.}
\usage{
distribute_MUs_in_muscle(muscle, MUs)
}
\arguments{
\item{muscle}{A single muscle object, must have element \code{MU_distributor}.}

\item{MUs}{A list of MU objects.}
}
\value{
A list of modified MU objects with assigned geometrical properties.
}
\description{
Distribute MUs throughout the base muscle's assigned area, taking care of
assumed fiber densities, muscle boundaries, etc. This is a purely technical
function; the algorithm used for MU placement must be provided externally via
\code{muscle$MU_distributor}.
}
\details{
Returned objects have one very important property: $transform, which
transforms coordinates relative to the MU base region (a (0,1)x(0,1) square)
into coordinates relative to the muscle's base region (also a [0,1]x[0,1]
square). It is a function that takes 2D vectors with values between 0 and 1
in both dimensions. The same is true for its return values.
}
