% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{distribute_MUs_reg}
\alias{distribute_MUs_reg}
\title{Distribute motor units on a regular grid across the muscle cross-section}
\usage{
distribute_MUs_reg(MUs, num_MUs_x, num_MUs_y, overlap = 0)
}
\arguments{
\item{MUs}{A list of MU objects. Must contain \code{num_MUs_x * num_MUs_y}
MUs.}

\item{num_MUs_x}{The number of MUs to be placed in one row in x direction.}

\item{num_MUs_y}{The number of MUs to be placed in one row in y direction.}

\item{overlap}{The amount of overlap to occur between neighboring MU regions.
Currently, only \code{overlap == 0} is implemented.}
}
\value{
A list of MU objects, containing the elements of \code{MUs} but with
  additional properties:
\itemize{
 \item{"$pos"}{The relative 2-D position (i.e. from [0,1] x [0,1]) of the MUs
               center on the muscle cross section.}
 \item{"$length_x"}{The relative size of the MU's surrounding rectangle box in
                  x direction on the muscle cross section. Value between 0
                  and 1. Used to transform coordinates relative to the MU
                  area to coordinates relative to the muscle area.}
 \item{"$length_y"}{The relative size of the MU's surrounding rectangle box in
                  y direction on the muscle cross section. Value between 0
                  and 1. Used to transform coordinates relative to the MU
                  area to coordinates relative to the muscle area.}
}
}
\description{
Deterministically create a regular grid of MU positions across the muscle
cross section for a muscle with a rectangular base shape. Assign the same
area
}
\details{
The MU areas are obtained by simple division of the unit square into a
(num_MUs_x x num_MUs_y) grid. No randomness is involved in the MU
distribution. All MUs are of the same size. Currently, only exact division of
the muscle area into MU areas is implemeted, i.e., no overlapping MUs and no
muscle sections without MUs.
}
