% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{generate_fibers}
\alias{generate_fibers}
\title{Generate muscle fibers for a motor unit}
\usage{
generate_fibers(muscle, MU)
}
\arguments{
\item{muscle}{A muscle object.}

\item{MU}{A MU object.}
}
\value{
A list of muscle fiber objects generated by applying the definitions
given in the muscle and MU arguments.
}
\description{
Take a muscle and a MU configuration and return a list of associated muscle
fibers for that MU.
}
\details{
Initializes the geometry by calling \code{muscle$rel_fiber_distributor(MU)}
(which should be defined in the config file) and then calculates all other
fiber properties from the MU's properties and the geometrical position of the
fiber.
}
