% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{generate_MUs}
\alias{generate_MUs}
\title{Generate motor units (MUs) and muscle fibers for all muscles}
\usage{
generate_MUs(muscles)
}
\arguments{
\item{muscles}{A data.frame containing information on the muscle properties.
Will usually be generated by a call to \link{convert_muscle_list_to_df}.}
}
\value{
A data.frame containing categorical columns 'muscle' and 'MU'
  giving the corresponding indices and 'MU.obj' containing the actual MU
  objects. These in turn have a property $fibers that contains a list of the
  muscle fibers contained in that MU. No geometrical information is present
  at this stage.
}
\description{
Set up the muscle structure by generating motor units (MUs) and their
contained muscle fibers according to the given configuration.
}
\details{
Generates a set of MUs parameters following an exponential distribution with
muscle-specific parameters and then gathers all the MUs generated for the
different muscles in a single data.frame.
}
