% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_muscles.r
\name{generate_MUs_for_muscle}
\alias{generate_MUs_for_muscle}
\title{Generate Motor Units for a Muscle}
\usage{
generate_MUs_for_muscle(muscle)
}
\arguments{
\item{muscle}{A muscle object.}
}
\value{
A list of \code{muscle$num_MUs} MU objects. Contains MU objects of
  class either "MU_rect" or "MU_circ", as defined by \code{muscle$MU_class}.
}
\description{
Assign MU properties according to muscle parameters.
}
\details{
MUs have a set of parameters (recruitment thresholds, peak twitch force,
number of innervated fibers) that are closely related, and which follow an
exponential distribution: There are many MUs that are generated early, have a
low peak twitch force and innervate few fibers. On the other side of the
spectrum, there are few MUs that are recruited only at high force demand
levels, attain a high peak twitch force and innervate many fibers.
The exact parameters of the distribution are a property of the muscle.
}
