% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{get_unique_idces}
\alias{get_unique_idces}
\title{Returns the indices of the first occurences of unique elements in a vector.}
\usage{
get_unique_idces(vec)
}
\description{
Returns the indices of the first occurences of unique elements in a vector.
}
\note{
This is _highly_ optimizable.
}
