% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_Rosenfalck.r
\name{IAP_Rosenfalck}
\alias{IAP_Rosenfalck}
\title{Rosenfalck's IAP model function}
\usage{
IAP_Rosenfalck(z, A = 96 * 0.001, B = -90 * 0.001)
}
\arguments{
\item{z}{The spatial variable. Can be a scalar or a vector. Should be in
\code{mm} when employing the default parameters for \code{A} and \code{B}.}

\item{A}{Model parameter. Must be a scalar value in [V/mm^3]. Default value
is 96 mv/mm^3.}

\item{B}{Model parameter. Must be a scalar value in SI units. Default value
is -90mV.}
}
\value{
A numerical vector containing the value(s) of Rosenfalcks function at
  position(s) z.
}
\description{
An implementation of a variant of Rosenfalck's model of the Intracellular
Action Potential (IAP) wave shape.  The model is simply
}
\details{
Vm(z[mm]) = Az^3 exp(-z) + B, if z > 0 and
   Vm(z[mm]) = B, if z <= 0.

Note that the front of the IAP is towards z = 0, whereas the tail of the wave
is around z = 15. The wave is running in negative z direction.
}
