% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_tools.r
\name{mirror_hermitian_fft}
\alias{mirror_hermitian_fft}
\title{Mirror a set of FFT values by assuming the fourier transform to be hermitian.}
\usage{
mirror_hermitian_fft(fft_vals, sampling)
}
\arguments{
\item{fft_vals}{A complex vector containing the FFT values of a function for
the frequencies given by sampling$freqs_to_calc (i.e. f=0 as well as the
positive ones).}

\item{sampling}{A sampling object, as usually created by
\code{create.sampling}.}
}
\value{
A complex vector containing the full set of FFT values at all
frequencies (and in the same order as) given by sampling$fft_freqs.
}
\description{
The FT of a purely real function is hermitian. Exploit this to obtain a full
set of FFT values from the values at positive frequencies only (and f=0).
}
\note{
Should be clear, but once again: This function produces false results
but \strong{does not throw an error} if the underlying function is not purely
real!
}
