% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.r
\name{plot_TFs}
\alias{plot_TFs}
\title{Old & currently broken. Keeping this only for possible future use.}
\usage{
plot_TFs(sim_result, MUs = 1:3, print_plot = TRUE)
}
\description{
Old & currently broken. Keeping this only for possible future use.
}
