% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAP_Rosenfalck.r
\name{psi_Rosenfalck}
\alias{psi_Rosenfalck}
\title{An implementation of Rosenfalck's model of the first derivative of the
Intracellular Action Potential (IAP) wave shape, evaluated in the negative
z direction. This quantity is called psi.}
\usage{
psi_Rosenfalck(z, A = 96 * 0.001, B = -90 * 0.001)
}
\arguments{
\item{z}{The spatial variable. Can be a scalar or a vector. Should be in
\code{m}.}

\item{A}{Model parameter. Must be a scalar value in [V/mm^3]. Default value
is 96 mv/mm^3.}

\item{B}{Model parameter. Must be a scalar value in SI units. Default value
is -90mV.}
}
\value{
A numerical vector containing the value(s) of Rosenfalcks psi
function (in [V]) at position(s) z [m].
}
\description{
psi(z[mm]) = (Vm(-z))'
= | -(z+3) A z^2 exp(z)                                if z < 0
| 0                                                  if z > 0
| -(z+3) A z^2 exp(z) + A z^3 exp(z) delta(z), if z = 0.
}
\details{
Here, we implement psi(0) = 0.
Note that the front of the IAP is towards z = 0, whereas the tail of the wave
is around z = -15. It is running in positive z direction.
}
