% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_coding.r
\name{rate_function}
\alias{rate_function}
\title{This implements an exponential model of the firing-rate-force relationship
relationship, as proposed in "Relationship between firing rate and
recruitment threshold of motoneurons in voluntary isometric contractions"
[De Luca, Hostage] (2010). The model is}
\usage{
rate_function(muscle_load, params)
}
\description{
fr_i(x) = Dx + (C - A exp(-x/B)) x_i + E
}
\details{
where fr_i(x) denotes the firing rate of motoneuron i at excitation level x,
and x_i denotes the recruitment threshold of that motoneuron. The remaining
parameters are constant parameters of the rate coding model that describe the
rate coding behavior of a particular muscle.

The model respects the following principles: \itemize{
\item The 'onion-skin' phenomenon: at each point in time, earlier recruited
MUs maintain higher firing rates than later recruited ones.
\item Firing-rate curves converge to similar (here: the same) firing rates at
high force levels.
\item At a given value of force, higher-threshold MUs display higher slopes
in their firing-rate-force curves.
\item Initial firing rates of MUs increase with the recruitment threshold.
}
}
