% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{rnorm_bounded}
\alias{rnorm_bounded}
\title{Helper function for random number generation}
\usage{
rnorm_bounded(min = 0, max = 1, n_sigma = 2)
}
\arguments{
\item{min}{The lower boundary of the interval to be sampled from.}

\item{max}{The upper boundary of the interval to be sampled from.}

\item{n_sigma}{An integer number indicating the range of the Gaussian
distribution be accepted without resampling.}
}
\value{
A single numerical value lying in the given interval.
}
\description{
Takes an intended range of values for the property and returns a randomly
instantiated value.
}
\details{
This function considers the given range \code{[min, max]} to be the
\code{n}-sigma range of a Gaussian probability distribution, with the mean
lying in the center of the range. Samples outside of this interval
(i.e. outside of the \code{n}-sigma range, hence e.g. less than 5\% of all
samples for \code{n = 2}) are rejected and resampled.
}
