% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{rotate2D}
\alias{rotate2D}
\title{Rotate a 2D vector by an angle}
\usage{
rotate2D(x, angle)
}
\arguments{
\item{x}{The two-component vector(s) that is (are) to be rotated. In case
of multiple vectors, they should be in the form of a 2xn array or matrix.}

\item{angle}{A single numeric specifying the angle by which to rotate about
the origin, in radians.}
}
\value{
A two-component vector that contains the rotated x, or a matrix
  containing the rotated columns of x.
}
\description{
Rotates a two-dimensional vector about the origin.
}
\author{
Eike Petersen
}
