% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{rotate3D_lh}
\alias{rotate3D_lh}
\title{Rotate a 3D vector around the three axes in a left-handed system}
\usage{
rotate3D_lh(x, angles)
}
\arguments{
\item{x}{The three-component vector(s) that is (are) to be rotated. In case
of multiple vectors, they should be in the form of a 3xn array or matrix.}

\item{angles}{A three-component vector that contains the angles by which to
rotate around the axes in radians and in the order x, y, z. Positive angles
rotate counter-clockwise in left-handed coordinate systems.}
}
\value{
A three-component vector that contains the rotated x, or a matrix
  containing the rotated columns of x.
}
\description{
Rotates a three-dimensional vector by the three angles specified around the
three axes of a left-handed coordinate system.
}
\details{
If more than one angle is non-zero, the rotations will be applied in the
order x-y-z.
}
\note{
Rotation is not as easily invertible as might seem apparent:
  \code{rotate3D_lh(rotate3D_lh(vec, angles), -angles) != vec}
  at least if there is more than one non-zero component in angles.
}
