% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{setup_muscles}
\alias{setup_muscles}
\title{Muscle setup}
\usage{
setup_muscles(config_file, deterministic = FALSE, out_file = NULL)
}
\description{
Muscle setup
}
\examples{
model <- setup_muscles("config/integration.r", deterministic = TRUE) 
}
