% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_simple_TFs.r
\name{simple.TF}
\alias{simple.TF}
\title{Calculation of simple MU - electrode transfer functions (inner wrapper)}
\usage{
simple.TF(.MU, .electrode, .kt)
}
\arguments{
\item{.MU}{The motor unit which is the source of the surface potential.}

\item{.electrode}{The electrode at which the potential is measured.}

\item{.kt}{A vector containing the discrete angular frequencies at which the
transfer function is to be evaluated.}
}
\value{
A list object providing two properties: First, a property
  '$sampled.data' containing the discrete values of the transfer function
  from motor unit .MU to electrode .electrode obtained by evaluation at the
  discrete angular frequencies kt. Values are _not_ normalized according to
  .dist.min. And second, a property '$amp' that contains the sum over all
  muscle fibers of the squared inverse of the distance from that fiber's
  innvervation zone to the electrode position. This will be used for
  normalization later on.
}
\description{
Compute discrete, unnormalized values of the transfer function from the motor
unit .MU to the electrode .electrode. For the transfer function model, see
comments above.
}
