% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.r
\name{simulate_MU_wrapper}
\alias{simulate_MU_wrapper}
\title{Simple worker function wrapper to facilitate progression display}
\usage{
simulate_MU_wrapper(x, p, ...)
}
\description{
Simple worker function wrapper to facilitate progression display
}
