% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_muscle}
\alias{simulate_muscle}
\title{Simulate a surface electromyogram and force resulting from muscle acticity}
\usage{
simulate_muscle(
  config_file,
  deterministic = FALSE,
  out_file = NULL,
  num_cores = 1
)
}
\arguments{
\item{config_file}{Specifies the simulation config file to be read. Must
specify a number of physiological parameters. See section Details for more
information.}

\item{out_file}{Optional. If given, specifies a file to save the simulation
\code{environment()}. Results can be loaded via \code{readRDS(out_file)}
afterwards.}

\item{num_cores}{Number of cores to use for various parts of the simulatino
that can be performed in parallel. Optional, default=1.}

\item{MoreArgs}{A list of parameter name - value pairs. Can be used to
provide necessary model parameters that are missing in the
\code{config_file}.}
}
\description{
Given a virtual experimental set-up, simulate the sEMG and force data
resulting at a set of electrodes from activity in several muscles.
}
\details{
Implements a physiologically sound and detailed mathematical model of motor
unit activity and the volume conductors represented by skin and fat tissues
between muscle and electrodes. Many (in fact, practically all except the
underlying base model) physiological parameters can be configured by
providing an appropriate config file.

For more details regarding the simulation model, refer to the paper
Petersen and Rostalski (2019): "A Comprehensive Mathematical Model of Motor
Unit Pool Organization, Surface Electromyography and Force Generation",
doi: 10.3389/fphys.2019.00176.
}
\note{
Currently, effects resulting from fatigue are not implemented.
}
\author{
Eike Petersen
}
