% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_firing_contribs.r
\name{sum_muscle_contribs}
\alias{sum_muscle_contribs}
\title{Combine the contributions of all individual MUs to each electrode to obtain the final electrode signals.}
\usage{
sum_muscle_contribs(muscle_firing_contribs)
}
\description{
Combine the contributions of all individual MUs to each electrode to obtain the final electrode signals.
}
