% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_muscles.r
\name{summarize_muscles}
\alias{summarize_muscles}
\title{Summarize key properties of generated muscles to command line}
\usage{
summarize_muscles(sim_result, muscles = NA)
}
\description{
Summarize key properties of generated muscles to command line
}
