% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_muscle}
\alias{simulate_muscle}
\title{Simulate a surface electromyogram and force resulting from muscle activity}
\usage{
simulate_muscle(
  config_file,
  deterministic = FALSE,
  out_file = NULL,
  num_cores = future::availableCores(omit = 1)
)
}
\arguments{
\item{config_file}{Specifies the simulation config file to be read.
Must specify a number of physiological parameters including the
force function for each muscle. See section Details for more information.}

\item{deterministic}{Option sets seed for reproducible random
numbers.}

\item{out_file}{Optional. If given, specifies a file to save
the simulation \code{environment()}. Results can be loaded via
\code{readRDS(out_file)} afterwards.}

\item{num_cores}{Optional. Number of cores to use for various
parts of the simulation that can be performed in parallel.
Default = all available cores minus 1.}
}
\description{
Given a virtual experimental set-up, simulate the sEMG and force data
resulting at a set of electrodes from activity in several muscles.
}
\details{
Implements a physiologically sound and detailed mathematical model of motor
unit activity and the volume conductors represented by skin and fat tissues
between muscle and electrodes. Many (in fact, practically all except the
underlying base model) physiological parameters can be configured by
providing an appropriate config file.

For more details regarding the simulation model, refer to the paper
Petersen and Rostalski (2019): "A Comprehensive Mathematical Model of Motor
Unit Pool Organization, Surface Electromyography and Force Generation",
doi: 10.3389/fphys.2019.00176.
}
\note{
Currently, effects resulting from fatigue are not implemented.
}
\examples{
model <- simulate_muscle("config/abdominals.r", deterministic=TRUE, num_cores=3, out_file = "simout/abdominals.RDS")
model <- simulate_muscle("config/integration.r", deterministic = TRUE) 
}
\author{
Eike Petersen
}
